grades can be saved and loaded from the server now!
made the grade history set the date and time on the server side

Also made doubles be casted to floats correctly (mongo…)


Former-commit-id: 3254f1c467b537c021980f874722ee06f4bbb709

+++ CourseSketchProjects/CourseSketchDatabaseServer/src/main/java/database/institution/mongo/GradeManager.java
@@ -443,7 +443,7 @@ public final class GradeManager {
         }
 
         if (grade.containsField(CURRENT_GRADE)) {
-            protoGrade.setCurrentGrade((float) grade.get(CURRENT_GRADE));
+            protoGrade.setCurrentGrade((float) ((double) grade.get(CURRENT_GRADE)));
         }
 
         if (grade.containsField(GRADE_HISTORY)) {
@@ -474,7 +474,7 @@ public final class GradeManager {
     static GradeHistory buildProtoGradeHistory(final DBObject history) throws DatabaseAccessException {
         final GradeHistory.Builder protoHistory = GradeHistory.newBuilder();
         if (history.containsField(GRADE_VALUE)) {
-            protoHistory.setGradeValue((float) history.get(GRADE_VALUE));
+            protoHistory.setGradeValue((float) ((double) history.get(GRADE_VALUE)));
         }
 
         if (history.containsField(COMMENT)) {
